Seperate layout for easier split in build time
