+++ client/layout-wrapper/logged-out.jsx
@@ -1,0 +1,14 @@
+/**
+ * External dependencies
+ */
+import React from 'react';
+
+/**
+ * Internal Dependencies
+ */
+import LayoutLoggedOut from 'layout/logged-out';
+
+const LoggedOutLayout = ( { primary, secondary, redirectUri } ) =>
+	<LayoutLoggedOut primary={ primary } secondary={ secondary } redirectUri={ redirectUri } />;
+
+export default LoggedOutLayout;
